# docs/contracts/account.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aisenhub Account API",
    "version": "1.0.0",
    "description": "Frozen V1 Account contract. Route implementations are delivered by later M2-M4 tasks; contract-only routes must not return fake success."
  },
  "servers": [{ "url": "https://account.example.invalid" }],
...
  "paths": {
...
    "/v1/auth/recent-proof": {
      "post": {
        "operationId": "issueRecentAuthProof",
        "tags": ["account"],
        "x-module": "M2",
        "x-implementation-status": "local-only",
        "x-reauthentication-event": "supabase-auth-reauthentication-otp",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "$ref": "#/components/requestBodies/Reauthentication"
        },
        "responses": {
          "201": { "$ref": "#/components/responses/RecentAuthProof" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "503": { "$ref": "#/components/responses/Unavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "platformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Key"
      }
    },
...
    "requestBodies": {
...
      "Reauthentication": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["token"],
              "properties": {
                "token": {
                  "type": "string",
                  "pattern": "^[0-9]{6}$"
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
    },
    "headers": {
      "NoStore": {
        "description": "V1 responses are not shared or ISR cached.",
        "schema": { "type": "string", "const": "no-store" }
      },
...
      "RequestId": { "schema": { "type": "string", "format": "uuid" } }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "$ref": "#/components/schemas/ApiErrorCode" },
          "message": { "type": "string" },
          "details": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          }
        },
        "additionalProperties": false
      },
      "ApiErrorCode": {
        "type": "string",
        "enum": [
          "INVALID_INPUT",
          "UPLOAD_SIZE_MISMATCH",
          "PAYLOAD_TOO_LARGE",
          "UNAUTHORIZED",
          "PLATFORM_CREDENTIAL_INVALID",
          "SESSION_REVOKED",
          "PLATFORM_DISABLED",
          "ACCOUNT_SUSPENDED",
          "ACCOUNT_CLOSED",
          "ADMIN_REQUIRED",
          "MFA_REQUIRED",
          "RECENT_MFA_REQUIRED",
          "GLOBAL_DELETE_PENDING",
          "ACCOUNT_NOT_ACTIVATED",
          "ACTIVATION_DISABLED",
          "PLAN_CONFLICT",
          "ENTITLEMENT_PERPETUAL",
          "ENTITLEMENT_SUSPENDED",
          "IDEMPOTENCY_CONFLICT",
          "OPERATION_IN_PROGRESS",
          "FILE_BUSY",
          "FILE_CONTENT_CONFLICT",
          "REPLACEMENT_CAPACITY_REQUIRED",
          "QUOTA_EXCEEDED",
          "RESOURCE_NOT_FOUND",
          "INVALID_CODE",
          "CODE_EXPIRED",
          "UPLOAD_INTENT_EXPIRED",
          "CODE_DISABLED",
          "CODE_ALREADY_REDEEMED",
          "PRECONDITION_FAILED",
          "PRECONDITION_REQUIRED",
          "RATE_LIMITED",
          "AUTHORIZATION_UNAVAILABLE",
          "STORAGE_UNAVAILABLE"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ApiError" },
          "request_id": { "type": "string", "format": "uuid" }
        },
        "additionalProperties": false
      },
...
      "RecentAuthProof": {
        "type": "object",
        "required": ["data", "request_id"],
        "properties": {
          "data": {
            "type": "object",
            "required": ["proof_id", "expires_at"],
            "properties": {
              "proof_id": { "type": "string", "format": "uuid" },
              "expires_at": { "type": "string", "format": "date-time" }
            },
            "additionalProperties": false
          },
          "request_id": { "type": "string", "format": "uuid" }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "RecentAuthProof": {
        "description": "Server-issued proof bound to the current user session for five minutes.",
        "headers": {
          "Cache-Control": { "$ref": "#/components/headers/NoStore" },
          "X-Request-Id": { "$ref": "#/components/headers/RequestId" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/RecentAuthProof" }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid input.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Authentication or platform credential failure.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
      "Forbidden": {
        "description": "Authorization or lifecycle refusal.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      },
...
      "Unavailable": {
        "description": "Authorization or Storage dependency unavailable.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    }
  }
}